[crypto] moved Ed25519 implementation in its own directory
