+++ crates/aptos-crypto/src/ed25519/ed25519_keys.rs
@@ -1,62 +1,26 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
 
-//! This module provides an API for the PureEdDSA signature scheme over the Ed25519 twisted
-//! Edwards curve as defined in [RFC8032](https://tools.ietf.org/html/rfc8032).
-//!
-//! Signature verification also checks and rejects non-canonical signatures.
-//!
-//! # Examples
-//!
-//! ```
-//! use aptos_crypto_derive::{CryptoHasher, BCSCryptoHash};
-//! use aptos_crypto::{
-//!     ed25519::*,
-//!     traits::{Signature, SigningKey, Uniform},
-//! };
-//! use rand::{rngs::StdRng, SeedableRng};
-//! use rand_core::OsRng;
-//! use serde::{Serialize, Deserialize};
-//! use aptos_crypto::test_utils::KeyPair;
-//!
-//! #[derive(Serialize, Deserialize, CryptoHasher, BCSCryptoHash)]
-//! pub struct TestCryptoDocTest(String);
-//! let message = TestCryptoDocTest("Test message".to_string());
-//!
-//! let mut rng = OsRng;
-//! let kp = KeyPair::<Ed25519PrivateKey, Ed25519PublicKey>::generate(&mut rng);
-//!
-//! let signature = kp.private_key.sign(&message);
-//! assert!(signature.verify(&message, &kp.public_key).is_ok());
-//! ```
-#![allow(clippy::integer_arithmetic)]
+//! This file implements traits for Ed25519 private keys and public keys.
 
 use crate::{
+    ed25519::{Ed25519Signature, ED25519_PRIVATE_KEY_LENGTH, ED25519_PUBLIC_KEY_LENGTH},
     hash::CryptoHash,
     traits::*,
 };
-use anyhow::{anyhow, Result};
 use aptos_crypto_derive::{DeserializeKey, SerializeKey, SilentDebug, SilentDisplay};
 use core::convert::TryFrom;
 use serde::Serialize;
-use std::{cmp::Ordering, fmt};
+use std::fmt;
 
-/// The length of the Ed25519PrivateKey
-pub const ED25519_PRIVATE_KEY_LENGTH: usize = ed25519_dalek::SECRET_KEY_LENGTH;
-/// The length of the Ed25519PublicKey
-pub const ED25519_PUBLIC_KEY_LENGTH: usize = ed25519_dalek::PUBLIC_KEY_LENGTH;
-/// The length of the Ed25519Signature
-pub const ED25519_SIGNATURE_LENGTH: usize = ed25519_dalek::SIGNATURE_LENGTH;
-
-/// The order of ed25519 as defined in [RFC8032](https://tools.ietf.org/html/rfc8032).
-const L: [u8; 32] = [
-    0xed, 0xd3, 0xf5, 0x5c, 0x1a, 0x63, 0x12, 0x58, 0xd6, 0x9c, 0xf7, 0xa2, 0xde, 0xf9, 0xde, 0x14,
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10,
-];
+#[cfg(any(test, feature = "fuzzing"))]
+use crate::test_utils::{self, KeyPair};
+#[cfg(any(test, feature = "fuzzing"))]
+use proptest::prelude::*;
 
 /// An Ed25519 private key
 #[derive(DeserializeKey, SerializeKey, SilentDebug, SilentDisplay)]
-pub struct Ed25519PrivateKey(ed25519_dalek::SecretKey);
+pub struct Ed25519PrivateKey(pub(crate) ed25519_dalek::SecretKey);
 
 #[cfg(feature = "assert-private-keys-not-cloneable")]
 static_assertions::assert_not_impl_any!(Ed25519PrivateKey: Clone);
@@ -71,11 +35,7 @@ impl Clone for Ed25519PrivateKey {
 
 /// An Ed25519 public key
 #[derive(DeserializeKey, Clone, SerializeKey)]
-pub struct Ed25519PublicKey(ed25519_dalek::PublicKey);
-
-/// An Ed25519 signature
-#[derive(DeserializeKey, Clone, SerializeKey)]
-pub struct Ed25519Signature(ed25519_dalek::Signature);
+pub struct Ed25519PublicKey(pub(crate) ed25519_dalek::PublicKey);
 
 impl Ed25519PrivateKey {
     /// The length of the Ed25519PrivateKey
@@ -162,73 +122,6 @@ impl Ed25519PublicKey {
             .to_edwards(sign)
             .ok_or(CryptoMaterialError::DeserializationError)?;
         Ed25519PublicKey::try_from(&ed_point.compress().as_bytes()[..])
-    }
-}
-
-impl Ed25519Signature {
-    /// The length of the Ed25519Signature
-    pub const LENGTH: usize = ed25519_dalek::SIGNATURE_LENGTH;
-
-    /// Serialize an Ed25519Signature.
-    pub fn to_bytes(&self) -> [u8; ED25519_SIGNATURE_LENGTH] {
-        self.0.to_bytes()
-    }
-
-    /// Deserialize an Ed25519Signature without any validation checks (malleability)
-    /// apart from expected signature size.
-    pub(crate) fn from_bytes_unchecked(
-        bytes: &[u8],
-    ) -> std::result::Result<Ed25519Signature, CryptoMaterialError> {
-        match ed25519_dalek::Signature::try_from(bytes) {
-            Ok(dalek_signature) => Ok(Ed25519Signature(dalek_signature)),
-            Err(_) => Err(CryptoMaterialError::DeserializationError),
-        }
-    }
-
-    /// return an all-zero signature (for test only)
-    #[cfg(any(test, feature = "fuzzing"))]
-    pub fn dummy_signature() -> Self {
-        Self::from_bytes_unchecked(&[0u8; Self::LENGTH]).unwrap()
-    }
-
-    /// Check for correct size and third-party based signature malleability issues.
-    /// This method is required to ensure that given a valid signature for some message under some
-    /// key, an attacker cannot produce another valid signature for the same message and key.
-    ///
-    /// According to [RFC8032](https://tools.ietf.org/html/rfc8032), signatures comprise elements
-    /// {R, S} and we should enforce that S is of canonical form (smaller than L, where L is the
-    /// order of edwards25519 curve group) to prevent signature malleability. Without this check,
-    /// one could add a multiple of L into S and still pass signature verification, resulting in
-    /// a distinct yet valid signature.
-    ///
-    /// This method does not check the R component of the signature, because R is hashed during
-    /// signing and verification to compute h = H(ENC(R) || ENC(A) || M), which means that a
-    /// third-party cannot modify R without being detected.
-    ///
-    /// Note: It's true that malicious signers can already produce varying signatures by
-    /// choosing a different nonce, so this method protects against malleability attacks performed
-    /// by a non-signer.
-    pub fn check_s_malleability(bytes: &[u8]) -> std::result::Result<(), CryptoMaterialError> {
-        if bytes.len() != ED25519_SIGNATURE_LENGTH {
-            return Err(CryptoMaterialError::WrongLengthError);
-        }
-        if !Ed25519Signature::check_s_lt_l(&bytes[32..]) {
-            return Err(CryptoMaterialError::CanonicalRepresentationError);
-        }
-        Ok(())
-    }
-
-    /// Check if S < L to capture invalid signatures.
-    fn check_s_lt_l(s: &[u8]) -> bool {
-        for i in (0..32).rev() {
-            match s[i].cmp(&L[i]) {
-                Ordering::Less => return true,
-                Ordering::Greater => return false,
-                _ => {}
-            }
-        }
-        // As this stage S == L which implies a non canonical S.
-        false
     }
 }
 
@@ -386,124 +279,17 @@ impl ValidCryptoMaterial for Ed25519PublicKey {
     }
 }
 
-//////////////////////
-// Signature Traits //
-//////////////////////
-
-impl Signature for Ed25519Signature {
-    type VerifyingKeyMaterial = Ed25519PublicKey;
-    type SigningKeyMaterial = Ed25519PrivateKey;
-
-    /// Verifies that the provided signature is valid for the provided message, going beyond the
-    /// [RFC8032](https://tools.ietf.org/html/rfc8032) specification, checking both scalar
-    /// malleability and point malleability (see documentation [here](https://docs.rs/ed25519-dalek/latest/ed25519_dalek/struct.PublicKey.html#on-the-multiple-sources-of-malleability-in-ed25519-signatures)).
-    ///
-    /// This _strict_ verification performs steps 1,2 and 3 from Section 5.1.7 in RFC8032, and an
-    /// additional scalar malleability check (via [Ed25519Signature::check_s_malleability][Ed25519Signature::check_s_malleability]).
-    ///
-    /// This function will ensure both the signature and the `public_key` are not in a small subgroup.
-    fn verify<T: CryptoHash + Serialize>(
-        &self,
-        message: &T,
-        public_key: &Ed25519PublicKey,
-    ) -> Result<()> {
-        Self::verify_arbitrary_msg(self, &signing_message(message), public_key)
-    }
-
-    /// Checks that `self` is valid for an arbitrary &[u8] `message` using `public_key`.
-    /// Outside of this crate, this particular function should only be used for native signature
-    /// verification in Move.
-    ///
-    /// This function will check both the signature and `public_key` for small subgroup attacks.
-    fn verify_arbitrary_msg(&self, message: &[u8], public_key: &Ed25519PublicKey) -> Result<()> {
-        // NOTE: ed25519::PublicKey::verify_strict already checks that the s-component of the signature
-        // is not mauled, but does so via an optimistic path which fails into a slower path. By doing
-        // our own (much faster) checking here, we can ensure dalek's optimistic path always succeeds
-        // and the slow path is never triggered.
-        Ed25519Signature::check_s_malleability(&self.to_bytes())?;
-
-        // NOTE: ed25519::PublicKey::verify_strict checks that the signature's R-component and
-        // the public key are *not* in a small subgroup.
-        public_key
-            .0
-            .verify_strict(message, &self.0)
-            .map_err(|e| anyhow!("{}", e))
-            .and(Ok(()))
-    }
-
-    fn to_bytes(&self) -> Vec<u8> {
-        self.0.to_bytes().to_vec()
-    }
-}
-
-impl Length for Ed25519Signature {
-    fn length(&self) -> usize {
-        ED25519_SIGNATURE_LENGTH
-    }
-}
-
-impl ValidCryptoMaterial for Ed25519Signature {
-    fn to_bytes(&self) -> Vec<u8> {
-        self.to_bytes().to_vec()
-    }
-}
-
-impl std::hash::Hash for Ed25519Signature {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        let encoded_signature = self.to_bytes();
-        state.write(&encoded_signature);
-    }
-}
-
-impl TryFrom<&[u8]> for Ed25519Signature {
-    type Error = CryptoMaterialError;
-
-    fn try_from(bytes: &[u8]) -> std::result::Result<Ed25519Signature, CryptoMaterialError> {
-        // We leave this check here to detect mauled signatures earlier, since it does not hurt
-        // performance much. (This check is performed again in Ed25519Signature::verify_arbitrary_msg
-        // and in ed25519-dalek's verify_strict API.)
-        Ed25519Signature::check_s_malleability(bytes)?;
-        Ed25519Signature::from_bytes_unchecked(bytes)
-    }
-}
-
-// Those are required by the implementation of hash above
-impl PartialEq for Ed25519Signature {
-    fn eq(&self, other: &Ed25519Signature) -> bool {
-        self.to_bytes()[..] == other.to_bytes()[..]
-    }
-}
-
-impl Eq for Ed25519Signature {}
-
-impl fmt::Display for Ed25519Signature {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", hex::encode(&self.0.to_bytes()[..]))
-    }
-}
-
-impl fmt::Debug for Ed25519Signature {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "Ed25519Signature({})", self)
-    }
-}
-
 /////////////
 // Fuzzing //
 /////////////
 
-#[cfg(any(test, feature = "fuzzing"))]
-use crate::test_utils::{self, KeyPair};
-
-/// Produces a uniformly random ed25519 keypair from a seed
+/// Produces a uniformly random Ed25519 keypair from a seed
 #[cfg(any(test, feature = "fuzzing"))]
 pub fn keypair_strategy() -> impl Strategy<Value = KeyPair<Ed25519PrivateKey, Ed25519PublicKey>> {
     test_utils::uniform_keypair_strategy::<Ed25519PrivateKey, Ed25519PublicKey>()
 }
 
-#[cfg(any(test, feature = "fuzzing"))]
-use proptest::prelude::*;
-
+/// Produces a uniformly random Ed25519 public key
 #[cfg(any(test, feature = "fuzzing"))]
 impl proptest::arbitrary::Arbitrary for Ed25519PublicKey {
     type Parameters = ();
